add error.blade.php,destroy.blade.php

+++ app/Http/Controllers/TasksController.php
@@ -161,18 +161,21 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        //$task->delete();
-        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
-        if (\Auth::id() === $task->user_id){
+        // もしuser_idが一致するユーザの場合指定されたidの$taskデータを削除する
+        if (\Auth::check() && \Auth::id() === $task->user_id){
+            // ログを出力して削除が実行されることを確認
+            \Log::info('Delete task: ' . $task->id);
+            // タスクを削除
             $task->delete();
-            //削除後はindexへリダイレクト
-            return redirect()->route('tasks.index')->with('success','Delete Successful');
+            //削除後は前のページへリダイレクト
+            return view('tasks.destroy');
+            //return back()->with('success','Delete Successful');
         }
 
-        // トップページへリダイレクトさせる
+        // 認証済みユーザではない場合はトップページへリダイレクトさせる
+        return view('tasks.error');
         //return redirect('/');
-        // 前のURLにリダイレクト
-        return back()->with('Delete Failed');
+        // 認証済みユーザではない場合は前のURLにリダイレクト
+        //return back()->with('Delete Failed');
     }
 }
